move types file up
